Reorganising chainimpl code
